Improved the code that determines the size of a call instruction. (This size is needed to place the return address in a register immediately before the call.)
For the X86 instruction set, it is virtually impossible to determine
the size of the opcode for a call instruction prior to emission of
the opcode. Therefore, the approach taken here splits up the basic
block after the call instruction. The address of the new basic block
that follows the call instruction is then used as a copy of the
return address for the call.